requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
Pillow>=10.1.0
django-mptt>=0.12.0